Remove sync mini cluster / harness
