+++ test/Knet.Kudu.Client.FunctionalTests/MiniCluster/MiniKuduCluster.cs
@@ -1,81 +1,89 @@
 ﻿using System;
+using System.Buffers;
 using System.Buffers.Binary;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Text;
+using System.Threading;
+using System.Threading.Tasks;
 using Knet.Kudu.Client.Connection;
-using Knet.Kudu.Client.Internal;
 using Knet.Kudu.Client.Protocol.Tools;
 using Knet.Kudu.Client.Util;
 using ProtoBuf;
 
 namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
 {
-    public class MiniKuduCluster : IDisposable
+    public class MiniKuduCluster : IAsyncDisposable
     {
         private readonly CreateClusterRequestPB _createClusterRequestPB;
-
-        // Control shell process.
-        private Process _miniCluster;
-
-        // Request channel to the control shell.
-        private Stream _miniClusterStdin;
-
-        // Response channel from the control shell.
-        private Stream _miniClusterStdout;
-
         private readonly Dictionary<HostAndPort, DaemonInfo> _masterServers;
         private readonly Dictionary<HostAndPort, DaemonInfo> _tabletServers;
+        private readonly SemaphoreSlim _singleRequest;
+
+        private ProcessEx _nativeProcess;
 
         public MiniKuduCluster(CreateClusterRequestPB createClusterRequestPB)
         {
             _createClusterRequestPB = createClusterRequestPB;
-
             _masterServers = new Dictionary<HostAndPort, DaemonInfo>();
             _tabletServers = new Dictionary<HostAndPort, DaemonInfo>();
+            _singleRequest = new SemaphoreSlim(1, 1);
         }
 
-        public void Start()
+        private Stream StdIn => _nativeProcess.StdIn;
+
+        private Stream StdOut => _nativeProcess.StdOut;
+
+        private Stream StdErr => _nativeProcess.StdErr;
+
+        public async Task StartAsync()
         {
             Directory.CreateDirectory(_createClusterRequestPB.ClusterRoot);
 
-            // TODO: Redirect standard error, and use async reads.
             var kuduExe = KuduBinaryLocator.FindBinary("kudu");
-            _miniCluster = new Process();
-            _miniCluster.StartInfo.WorkingDirectory = Path.GetDirectoryName(kuduExe);
-            _miniCluster.StartInfo.UseShellExecute = false;
-            _miniCluster.StartInfo.RedirectStandardInput = true;
-            _miniCluster.StartInfo.RedirectStandardOutput = true;
-            _miniCluster.StartInfo.FileName = kuduExe;
-            _miniCluster.StartInfo.ArgumentList.Add("test");
-            _miniCluster.StartInfo.ArgumentList.Add("mini_cluster");
-            _miniCluster.StartInfo.ArgumentList.Add("--serialization=pb");
+            var workingDirectory = Path.GetDirectoryName(kuduExe);
 
-            _miniCluster.Start();
+            var startInfo = new ProcessStartInfo
+            {
+                FileName = kuduExe,
+                WorkingDirectory = workingDirectory,
+                RedirectStandardInput = true,
+                RedirectStandardOutput = true,
+                RedirectStandardError = true,
+                UseShellExecute = false,
+                CreateNoWindow = true
+            };
 
-            _miniClusterStdin = _miniCluster.StandardInput.BaseStream;
-            _miniClusterStdout = _miniCluster.StandardOutput.BaseStream;
+            startInfo.ArgumentList.Add("test");
+            startInfo.ArgumentList.Add("mini_cluster");
+            startInfo.ArgumentList.Add("--serialization=pb");
+
+            _nativeProcess = new ProcessEx(startInfo);
+            _nativeProcess.Start();
+
+            // Start listening for events.
+            _ = ReadStdErrAsync();
 
             var createClusterRequest = new ControlShellRequestPB
             {
                 CreateCluster = _createClusterRequestPB
             };
-
-            SendRequestToCluster(createClusterRequest);
+            await SendRequestToClusterAsync(createClusterRequest);
 
             var startClusterRequest = new ControlShellRequestPB
             {
                 StartCluster = new StartClusterRequestPB()
             };
-            SendRequestToCluster(startClusterRequest);
+            await SendRequestToClusterAsync(startClusterRequest);
 
             // Initialize the maps of master and tablet servers.
             var getMastersRequest = new ControlShellRequestPB
             {
                 GetMasters = new GetMastersRequestPB()
             };
-            var masters = SendRequestToCluster(getMastersRequest);
+            var masters = await SendRequestToClusterAsync(getMastersRequest);
 
             foreach (var master in masters.GetMasters.Masters)
             {
@@ -88,7 +96,7 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
             {
                 GetTservers = new GetTServersRequestPB()
             };
-            var tablets = SendRequestToCluster(getTabletsRequest);
+            var tablets = await SendRequestToClusterAsync(getTabletsRequest);
 
             foreach (var tserver in tablets.GetTservers.Tservers)
             {
@@ -98,20 +106,15 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
             }
         }
 
-        public void Shutdown()
+        public async ValueTask DisposeAsync()
         {
-            _miniClusterStdin.Close();
-            _miniClusterStdout.Close();
-            _miniCluster.WaitForExit();
-        }
+            await StdIn.DisposeAsync();
+            await StdOut.DisposeAsync();
+            await StdErr.DisposeAsync();
 
-        public void Dispose()
-        {
-            Shutdown();
-
-            _miniClusterStdin.Dispose();
-            _miniClusterStdout.Dispose();
-            _miniCluster.Dispose();
+            await _nativeProcess.WaitUntilExitAsync();
+            _nativeProcess.Dispose();
+            _singleRequest.Dispose();
 
             try
             {
@@ -125,12 +128,17 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
             return KuduClient.NewBuilder(_masterServers.Keys.ToList()).Build();
         }
 
+        public KuduClientBuilder CreateClientBuilder()
+        {
+            return KuduClient.NewBuilder(_masterServers.Keys.ToList());
+        }
+
         /// <summary>
         /// Starts a master server identified by a host and port.
         /// Does nothing if the server was already running.
         /// </summary>
         /// <param name="hostPort">Unique host and port identifying the server.</param>
-        public void StartMasterServer(HostAndPort hostPort)
+        public async Task StartMasterServerAsync(HostAndPort hostPort)
         {
             var daemonInfo = GetMasterServer(hostPort);
             if (daemonInfo.IsRunning)
@@ -140,7 +148,7 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
             {
                 StartDaemon = new StartDaemonRequestPB { Id = daemonInfo.Id }
             };
-            SendRequestToCluster(request);
+            await SendRequestToClusterAsync(request);
 
             daemonInfo.IsRunning = true;
         }
@@ -150,7 +158,7 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
         /// Does nothing if the master was already dead.
         /// </summary>
         /// <param name="hostPort">Unique host and port identifying the server.</param>
-        public void KillMasterServer(HostAndPort hostPort)
+        public async Task KillMasterServerAsync(HostAndPort hostPort)
         {
             var daemonInfo = GetMasterServer(hostPort);
             if (!daemonInfo.IsRunning)
@@ -160,7 +168,7 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
             {
                 StopDaemon = new StopDaemonRequestPB { Id = daemonInfo.Id }
             };
-            SendRequestToCluster(request);
+            await SendRequestToClusterAsync(request);
 
             daemonInfo.IsRunning = false;
         }
@@ -170,7 +178,7 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
         /// Does nothing if the server was already running.
         /// </summary>
         /// <param name="hostPort">Unique host and port identifying the server.</param>
-        public void StartTabletServer(HostAndPort hostPort)
+        public async Task StartTabletServerAsync(HostAndPort hostPort)
         {
             var daemonInfo = GetTabletServer(hostPort);
             if (daemonInfo.IsRunning)
@@ -180,7 +188,7 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
             {
                 StartDaemon = new StartDaemonRequestPB { Id = daemonInfo.Id }
             };
-            SendRequestToCluster(request);
+            await SendRequestToClusterAsync(request);
 
             daemonInfo.IsRunning = true;
         }
@@ -190,7 +198,7 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
         /// Does nothing if the server was already dead.
         /// </summary>
         /// <param name="hostPort">Unique host and port identifying the server.</param>
-        public void KillTabletServer(HostAndPort hostPort)
+        public async Task KillTabletServerAsync(HostAndPort hostPort)
         {
             var daemonInfo = GetTabletServer(hostPort);
             if (!daemonInfo.IsRunning)
@@ -200,7 +208,7 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
             {
                 StopDaemon = new StopDaemonRequestPB { Id = daemonInfo.Id }
             };
-            SendRequestToCluster(request);
+            await SendRequestToClusterAsync(request);
 
             daemonInfo.IsRunning = false;
         }
@@ -209,40 +217,40 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
         /// Starts all the master servers.
         /// Does nothing to the servers that are already running.
         /// </summary>
-        public void StartAllMasterServers()
+        public async Task StartAllMasterServersAsync()
         {
             foreach (var hostPort in _masterServers.Keys)
-                StartMasterServer(hostPort);
+                await StartMasterServerAsync(hostPort);
         }
 
         /// <summary>
         /// Kills all the master servers.
         /// Does nothing to the servers that are already dead.
         /// </summary>
-        public void KillAllMasterServers()
+        public async Task KillAllMasterServersAsync()
         {
             foreach (var hostPort in _masterServers.Keys)
-                KillMasterServer(hostPort);
+                await KillMasterServerAsync(hostPort);
         }
 
         /// <summary>
         /// Starts all the tablet servers.
         /// Does nothing to the servers that are already running.
         /// </summary>
-        public void StartAllTabletServers()
+        public async Task StartAllTabletServersAsync()
         {
             foreach (var hostPort in _tabletServers.Keys)
-                StartTabletServer(hostPort);
+                await StartTabletServerAsync(hostPort);
         }
 
         /// <summary>
         /// Kills all the tablet servers.
         /// Does nothing to the servers that are already dead.
         /// </summary>
-        public void KillAllTabletServers()
+        public async Task KillAllTabletServersAsync()
         {
             foreach (var hostPort in _tabletServers.Keys)
-                KillTabletServer(hostPort);
+                await KillTabletServerAsync(hostPort);
         }
 
         // TODO: Set daemon flags (when Kudu 1.11 proto files are generated).
@@ -271,52 +279,75 @@ namespace Knet.Kudu.Client.FunctionalTests.MiniCluster
             throw new Exception($"Tablet server {hostPort} not found");
         }
 
-        private ControlShellResponsePB SendRequestToCluster(ControlShellRequestPB req)
+        private Task<ControlShellResponsePB> SendRequestToClusterAsync(ControlShellRequestPB req)
         {
-            using (var stream = new RecyclableMemoryStream())
+            var writer = new ArrayBufferWriter<byte>();
+            Serializer.Serialize(writer, req);
+            var messageLength = writer.WrittenCount;
+
+            var finalWriter = new ArrayBufferWriter<byte>();
+            var lengthSpan = finalWriter.GetSpan(4);
+            BinaryPrimitives.WriteInt32BigEndian(lengthSpan, messageLength);
+            finalWriter.Advance(4);
+
+            var span = finalWriter.GetSpan(messageLength);
+            writer.WrittenSpan.CopyTo(span);
+            finalWriter.Advance(messageLength);
+
+            var request = finalWriter.WrittenMemory;
+
+            return SendReceiveAsync(request);
+        }
+
+        private async Task<ControlShellResponsePB> SendReceiveAsync(ReadOnlyMemory<byte> request)
+        {
+            await _singleRequest.WaitAsync();
+
+            try
             {
-                // Make space to write the length of the entire message.
-                stream.GetMemory(4);
+                await StdIn.WriteAsync(request);
+                await StdIn.FlushAsync();
 
-                //Serializer.SerializeWithLengthPrefix(stream, req, PrefixStyle.Base128);
-                Serializer.Serialize(stream, req);
+                var buffer = new byte[4];
+                await ReadExactAsync(buffer);
+                var messageLength = BinaryPrimitives.ReadInt32BigEndian(buffer);
 
-                // Go back and write the length of the entire message, minus the 4
-                // bytes we already allocated to store the length.
-                BinaryPrimitives.WriteUInt32BigEndian(stream.AsSpan(), (uint)stream.Length - 4);
+                buffer = new byte[messageLength];
+                await ReadExactAsync(buffer);
 
-                _miniClusterStdin.Write(stream.AsSpan());
-                _miniClusterStdin.Flush();
+                var ms = new MemoryStream(buffer);
+                var response = Serializer.Deserialize<ControlShellResponsePB>(ms);
+
+                if (response.Error != null)
+                    throw new IOException(response.Error.Message);
+
+                return response;
             }
-
-            return Receive();
+            finally
+            {
+                _singleRequest.Release();
+            }
         }
 
-        private ControlShellResponsePB Receive()
-        {
-            var buffer = new byte[4];
-            ReadExact(buffer);
-            var messageLength = BinaryPrimitives.ReadInt32BigEndian(buffer);
-
-            buffer = new byte[messageLength];
-            ReadExact(buffer);
-
-            var ms = new MemoryStream(buffer);
-            var response = Serializer.Deserialize<ControlShellResponsePB>(ms);
-
-            if (response.Error != null)
-                throw new IOException(response.Error.Message);
-
-            return response;
-        }
-
-        private void ReadExact(Span<byte> buffer)
+        private async Task ReadExactAsync(Memory<byte> buffer)
         {
             do
             {
-                var read = _miniClusterStdout.Read(buffer);
+                var read = await StdOut.ReadAsync(buffer);
                 buffer = buffer.Slice(read);
             } while (buffer.Length > 0);
+        }
+
+        private async Task ReadStdErrAsync()
+        {
+            var buffer = new byte[4096];
+
+            while (true)
+            {
+                int read = await StdErr.ReadAsync(buffer);
+                var text = Encoding.UTF8.GetString(buffer, 0, read);
+                Console.Error.WriteLine(text);
+            }
         }
 
         private sealed class DaemonInfo
