+++ test/Knet.Kudu.Client.FunctionalTests/ScanPredicateTests.cs
@@ -12,7 +12,7 @@ namespace Knet.Kudu.Client.FunctionalTests
     [MiniKuduClusterTest]
     public class ScanPredicateTests : IAsyncLifetime
     {
-        private AsyncKuduTestHarness _harness;
+        private KuduTestHarness _harness;
         private KuduClient _client;
         private IKuduSession _session;
